+++ test/jest/__mock__/index.js
@@ -1,9 +1,0 @@
-import './createObjectURL.mock';
-import './currencyData.mock';
-import './matchMedia.mock';
-import './stripesConfig.mock';
-import './stripesCore.mock';
-import './stripesIcon.mock';
-import './stripesSmartComponents.mock';
-import './reactRouterDom.mock';
-import './reactQuery.mock';