+++ test/jest/interactors/index.js
@@ -1,3 +1,0 @@
-export { default as HeadlineInteractor } from './headline';
-export { default as IconButtonInteractor } from './icon-button';
-export { default as DatepickerInteractor } from './datepicker';