+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change history for stripes-erm-components
 
-## 7.1.0 In progress
-
+## 8.0.0 In progress
+* Removed all testing configuration and exports
+* Removed DEPRECATED Components
+  * LoadingPane
+  * withAsyncValidation
+  * withKiwtFieldArray
 ## 7.0.0 2022-10-26
 * ERM-2390 Licenses fails to add internal contact
 * ERM-2366 Core documents for a License shouldn't display a category